+++ pkg/local_object_storage/blobstor/fstree/range_bench_test.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"testing"
 
+	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
 	"github.com/stretchr/testify/require"
 )
 
@@ -48,7 +49,7 @@ func BenchmarkFSTree_GetRangeStream(b *testing.B) {
 					require.NoError(b, fsTree.Put(addr, obj.Marshal()))
 
 					for b.Loop() {
-						_, stream, err := fsTree.GetRangeStream(addr, tc.from, tc.length)
+						_, _, stream, err := fsTree.GetRangeStream(addr, common.NewPayloadRange(tc.from, tc.length), false)
 						if err == nil {
 							_, err = io.ReadFull(stream, buf)
 						}
@@ -64,7 +65,7 @@ func BenchmarkFSTree_GetRangeStream(b *testing.B) {
 
 					b.ResetTimer()
 					for k := range b.N {
-						_, stream, err := fsTree.GetRangeStream(addrs[k%len(addrs)], tc.from, tc.length)
+						_, _, stream, err := fsTree.GetRangeStream(addrs[k%len(addrs)], common.NewPayloadRange(tc.from, tc.length), false)
 						if err == nil {
 							_, err = io.ReadFull(stream, buf)
 						}