+++ pkg/services/object/get/service_test.go
@@ -9,6 +9,7 @@ import (
 	"sync/atomic"
 
 	iec "github.com/nspcc-dev/neofs-node/internal/ec"
+	blobcommon "github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
 	"github.com/nspcc-dev/neofs-node/pkg/services/object/common"
 	"github.com/nspcc-dev/neofs-node/pkg/services/object/util"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
@@ -120,6 +121,7 @@ type getECPartKey struct {
 
 type getECPartValue struct {
 	hdr object.Object
+	pld []byte
 	rdr io.ReadCloser
 	err error
 }
@@ -127,6 +129,31 @@ type getECPartValue struct {
 type mockLocalObjects struct {
 	unimplementedLocalStorage
 	getECPart map[getECPartKey]getECPartValue
+}
+
+func (x *mockLocalObjects) GetECPartRange(_ context.Context, cnr cid.ID, parent oid.ID, pi iec.PartInfo, rng blobcommon.PayloadRange, readHeader bool) (*object.Object, uint64, io.ReadCloser, error) {
+	v, ok := x.getECPart[getECPartKey{cnr: cnr, parent: parent, pi: pi}]
+	if !ok {
+		return nil, 0, nil, errors.New("[test] unexpected object requested")
+	}
+	if v.err != nil {
+		return nil, 0, nil, v.err
+	}
+
+	off, ln, err := rng.Resolve(uint64(len(v.pld)))
+	if err != nil {
+		return nil, 0, nil, err
+	}
+
+	var hdr *object.Object
+	if readHeader {
+		hdr = v.hdr.CutPayload()
+	}
+	if len(v.pld) == 0 {
+		return hdr, 0, nil, nil
+	}
+
+	return hdr, uint64(len(v.pld)), io.NopCloser(bytes.NewReader(v.pld[off:][:ln])), nil
 }
 
 func (x *mockLocalObjects) GetECPart(_ context.Context, cnr cid.ID, parent oid.ID, pi iec.PartInfo, allowAnyPart bool) (object.Object, io.ReadCloser, error) {
@@ -175,7 +202,7 @@ func (unimplementedObjectWriter) WriteChunk([]byte) error {
 
 type unimplementedLocalStorage struct{}
 
-func (x unimplementedLocalStorage) GetECPartRange(_ context.Context, _ cid.ID, _ oid.ID, _ iec.PartInfo, _, _ uint64) (uint64, io.ReadCloser, error) {
+func (x unimplementedLocalStorage) GetECPartRange(_ context.Context, _ cid.ID, _ oid.ID, _ iec.PartInfo, _ blobcommon.PayloadRange, _ bool) (*object.Object, uint64, io.ReadCloser, error) {
 	panic("unimplemented")
 }